requirements.txt
kivy
pillow
gTTS
moviepy
pydub
requests
